Rename Api namespace
